content/media/video/test/test_seek2.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: seek test 2</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video id='v'></video>
<pre id="test">
<script class="testbody" type="text/javascript">
// http://www.whatwg.org/specs/web-apps/current-work/#seek
// If the media element's networkState is less than LOADED_METADATA, 
// then the user agent must raise an INVALID_STATE_ERR exception 
var v = document.getElementById('v');
var passed = false;

ok(v.networkState < HTMLMediaElement.LOADED_METADATA,
   "Invalid network state in media element");

try {
  v.seek(1);
}
catch(e) {
  passed = true;
}

ok(passed, "Video did not raise error during invalid seek");
</script>
</pre>
</body>
</html>
